feat(domains): Redirects as a Web Domain tab (GH #1543)
Extract the redirects form body (whole-domain + per-page redirects, the
dnd-kit reorder list) from DomainRedirectsButton into a standalone
DomainRedirectsPanel, so it renders both inline as a tab on the tenant Web
Domain page and inside the row-menu Modal, which now delegates with
footer={null}.

The panel keeps the GH #717 clear semantics verbatim (an empty string, not
null, clears the whole-domain redirect) and re-syncs from the source domain on
change, keyed on JSON.stringify(page_redirects) so an unsaved edit survives a
background refetch. Saving invalidates both the domain list and the
single-domain query.

The tenant row menu is unchanged; stripping it down to Disable/Delete (with
the now-dead tenant-only modal wrappers removed) and folding DNS into a tab
are the next two slices.
